remove testing examples
