+++ examples/with-spark-schnorr/package.json
@@ -4,8 +4,6 @@
   "version": "0.1.0",
   "scripts": {
     "start": "tsx src/index.ts",
-    "auth-with-turnkey": "tsx src/authenticateWithTurnkey.ts",
-    "transfer-tokens": "tsx src/transferTokens.ts",
     "clean": "rimraf ./dist ./.cache",
     "typecheck": "tsc --noEmit"
   },
