+++ examples/with-spark-schnorr/src/index.ts
@@ -1,264 +1,187 @@
+/**
+ * Mints a new Spark token using Turnkey as the key custodian, then
+ * transfers the minted tokens to a receiver address.
+ *
+ * Flow:
+ *   1. Authenticate to SO (Turnkey signs the challenge via ECDSA/Schnorr)
+ *   2. CREATE transaction — announces the token (name, ticker, supply)
+ *   3. MINT transaction  — issues tokens to the issuer's own address
+ *   4. TRANSFER transaction — sends tokens to RECEIVER_SPARK_ADDRESS
+ *
+ * IssuerSparkWallet handles ALL transaction construction internally:
+ *   - Correct version numbers
+ *   - Fetching SO operator public keys
+ *   - Protobuf serialization / expiry time
+ *   - Hashing (hashPartialTokenTransaction)
+ *   - Broadcasting via gRPC
+ *
+ * Required env vars (in .env.local):
+ *   BASE_URL, API_PUBLIC_KEY, API_PRIVATE_KEY, ORGANIZATION_ID
+ *   TURNKEY_IDENTITY_ADDRESS, IDENTITY_PUBLIC_KEY_HEX
+ *   RECEIVER_SPARK_ADDRESS
+ *   SPARK_NETWORK   – REGTEST (default) or MAINNET
+ *
+ * Optional:
+ *   TOKEN_NAME      – default "TurnkeyTestToken"
+ *   TOKEN_TICKER    – default "TKT"
+ *   TOKEN_DECIMALS  – default 0
+ *   TOKEN_SUPPLY    – total max supply to create (default 1000000)
+ *   MINT_AMOUNT     – how many tokens to mint (default TOKEN_SUPPLY)
+ *   TRANSFER_AMOUNT – how many to transfer (default MINT_AMOUNT)
+ */
+
 import * as path from "path";
 import * as dotenv from "dotenv";
 dotenv.config({ path: path.resolve(process.cwd(), ".env.local") });
 
-import {
-  DefaultSparkSigner,
-  SparkWallet,
-  WalletConfig,
-  getSigHashFromTx,
-  getP2TRAddressFromPublicKey,
-  Network,
-} from "@buildonspark/spark-sdk";
-import { Transaction, p2tr, TEST_NETWORK } from "@scure/btc-signer";
-import { schnorr } from "@noble/curves/secp256k1";
-import { sha256 } from "@noble/hashes/sha2";
+import { IssuerSparkWallet } from "@buildonspark/issuer-sdk";
 import { Turnkey as TurnkeyServerSDK } from "@turnkey/sdk-server";
+import { TurnkeySparkSigner } from "./turnkeySigner";
 
 type SparkNetwork = "MAINNET" | "REGTEST";
 
-// ---------------------------------------------------------------------------
-// helpers
-// ---------------------------------------------------------------------------
+async function main() {
+  const network = (process.env.SPARK_NETWORK ?? "REGTEST") as SparkNetwork;
 
-function hex(bytes: Uint8Array): string {
-  return Buffer.from(bytes).toString("hex");
-}
+  const requiredVars = [
+    "API_PUBLIC_KEY",
+    "API_PRIVATE_KEY",
+    "ORGANIZATION_ID",
+    "TURNKEY_IDENTITY_ADDRESS",
+    "IDENTITY_PUBLIC_KEY_HEX",
+    "RECEIVER_SPARK_ADDRESS",
+  ] as const;
 
-// ---------------------------------------------------------------------------
-// 1. Initialize DefaultSparkSigner from mnemonic (low-level access to keys)
-// ---------------------------------------------------------------------------
-
-async function initSigner(mnemonic: string): Promise<DefaultSparkSigner> {
-  const signer = new DefaultSparkSigner();
-  const seed = await signer.mnemonicToSeed(mnemonic);
-  await signer.createSparkWalletFromSeed(seed);
-  return signer;
-}
-
-// ---------------------------------------------------------------------------
-// 2. Raw Schnorr signing comparison
-//    Signs an arbitrary payload with the Spark identity key (Schnorr) and,
-//    if Turnkey credentials are configured, compares against Turnkey's output.
-// ---------------------------------------------------------------------------
-
-async function compareSchnorrSignatures(
-  signer: DefaultSparkSigner,
-  payload: Uint8Array
-): Promise<void> {
-  console.log("\n── Schnorr signature (identity key) ──────────────────────────");
-
-  // noble's schnorr.sign does NOT hash internally — it expects a 32-byte msg.
-  // Pre-hash to 32 bytes so both Spark and Turnkey sign the same thing.
-  const msgHash = sha256(payload);
-  console.log(`Payload (hex):   ${hex(payload)}`);
-  console.log(`SHA-256 (hex):   ${hex(msgHash)}`);
-
-  const identityPubKey = await signer.getIdentityPublicKey();
-  // x-only (32-byte) pubkey required by BIP-340 schnorr.verify
-  const xOnlyPubKey = identityPubKey.slice(1);
-
-  const sparkSig = await signer.signSchnorrWithIdentityKey(msgHash);
-  console.log(`Spark SDK sig:   ${hex(sparkSig)}`);
-  const sparkValid = schnorr.verify(sparkSig, msgHash, xOnlyPubKey);
-  console.log(`Spark sig valid: ${sparkValid ? "✅  YES" : "❌  NO"}`);
-
-  const turnkeyAddress = process.env.TURNKEY_IDENTITY_ADDRESS;
-  if (
-    !process.env.API_PRIVATE_KEY ||
-    !process.env.API_PUBLIC_KEY ||
-    !process.env.ORGANIZATION_ID ||
-    !turnkeyAddress
-  ) {
-    console.log(
-      "\nSkipping Turnkey comparison – set TURNKEY_IDENTITY_ADDRESS and Turnkey credentials in .env.local to enable."
-    );
-    return;
+  for (const v of requiredVars) {
+    if (!process.env[v]) throw new Error(`Missing required env var: ${v}`);
   }
 
-  const client = new TurnkeyServerSDK({
+  const tokenName = process.env.TOKEN_NAME ?? "TurnkeyTestToken";
+  const tokenTicker = process.env.TOKEN_TICKER ?? "TKT";
+  const tokenDecimals = Number(process.env.TOKEN_DECIMALS ?? "0");
+  const maxSupply = BigInt(process.env.TOKEN_SUPPLY ?? "1000000");
+  const mintAmount = BigInt(process.env.MINT_AMOUNT ?? String(maxSupply));
+  const transferAmount = BigInt(
+    process.env.TRANSFER_AMOUNT ?? String(mintAmount)
+  );
+  const receiverSparkAddress = process.env.RECEIVER_SPARK_ADDRESS!;
+
+  // Initialize Turnkey client
+  const turnkeyClient = new TurnkeyServerSDK({
     apiBaseUrl: process.env.BASE_URL ?? "https://api.turnkey.com",
     apiPrivateKey: process.env.API_PRIVATE_KEY!,
     apiPublicKey: process.env.API_PUBLIC_KEY!,
     defaultOrganizationId: process.env.ORGANIZATION_ID!,
   });
 
-  // Send the pre-hashed 32-byte msg with NO_OP so Turnkey signs it as-is,
-  // matching exactly what Spark's signSchnorrWithIdentityKey received.
-  const { r, s } = await client.apiClient().signRawPayload({
-    signWith: turnkeyAddress,
-    encoding: "PAYLOAD_ENCODING_HEXADECIMAL",
-    hashFunction: "HASH_FUNCTION_NO_OP",
-    payload: hex(msgHash),
-  });
-
-  // Ensure r and s are each zero-padded to 32 bytes (64 hex chars)
-  const rPadded = r.padStart(64, "0");
-  const sPadded = s.padStart(64, "0");
-  const turnkeySigBytes = Buffer.from(rPadded + sPadded, "hex");
-  console.log(`Turnkey sig:     ${hex(turnkeySigBytes)}`);
-  const turnkeyValid = schnorr.verify(turnkeySigBytes, msgHash, xOnlyPubKey);
-  console.log(`Turnkey sig valid: ${turnkeyValid ? "✅  YES" : "❌  NO"}`);
-}
-
-// ---------------------------------------------------------------------------
-// 3. Bitcoin P2TR transaction signing (deposit key) on regtest
-//    Uses @scure/btc-signer + Spark SDK's signTransactionIndex.
-//    Requires UTXO_TXID, UTXO_VOUT, UTXO_VALUE, DESTINATION_ADDRESS in env.
-// ---------------------------------------------------------------------------
-
-async function signBitcoinTransaction(
-  signer: DefaultSparkSigner,
-  depositPubKey: Uint8Array
-): Promise<string> {
-  const txid = process.env.UTXO_TXID!;
-  const vout = parseInt(process.env.UTXO_VOUT!, 10);
-  const value = BigInt(process.env.UTXO_VALUE!);
-  const destination = process.env.DESTINATION_ADDRESS!;
-  const fee = BigInt(process.env.FEE_SATS ?? "300");
-
-  // x-only (32-byte) pubkey for taproot
-  const xOnly =
-    depositPubKey.length === 33 ? depositPubKey.slice(1) : depositPubKey;
-
-  const payment = p2tr(xOnly, undefined, TEST_NETWORK);
-
-  const tx = new Transaction();
-  tx.addInput({
-    txid,
-    index: vout,
-    witnessUtxo: {
-      script: payment.script,
-      amount: value,
-    },
-    tapInternalKey: xOnly,
-  });
-
-  tx.addOutputAddress(destination, value - fee, TEST_NETWORK);
-
-  // Compute the sighash before signing so we can log / compare it
-  const prevOut = {
-    script: payment.script,
-    amount: value,
-  };
-  const sighash = getSigHashFromTx(tx, 0, prevOut);
-  console.log("\n── Bitcoin P2TR transaction signing (deposit key) ────────────");
-  console.log(`Sighash (hex): ${hex(sighash)}`);
-
-  // Sign in place using the Spark SDK (Schnorr, deposit key)
-  signer.signTransactionIndex(tx, 0, depositPubKey);
-
-  tx.finalize();
-  const txHex = Buffer.from(tx.extract()).toString("hex");
-  console.log(`\nSigned tx hex:\n  ${txHex}`);
-  console.log(
-    "\nTo broadcast on regtest:\n  bitcoin-cli -regtest sendrawtransaction <hex>"
+  // Create the Turnkey-backed signer
+  const signer = new TurnkeySparkSigner(
+    turnkeyClient,
+    process.env.TURNKEY_IDENTITY_ADDRESS!,
+    process.env.IDENTITY_PUBLIC_KEY_HEX!
   );
-  return txHex;
-}
 
-// ---------------------------------------------------------------------------
-// 4. Full Spark wallet on REGTEST – tests that the SDK signs correctly
-//    enough for the Spark network to accept the transaction
-// ---------------------------------------------------------------------------
+  // Initialize IssuerSparkWallet with Turnkey signer.
+  // Cast to `any` is safe: the mismatched methods (subtractAndSplitSecretWithProofsGivenDerivations,
+  // etc.) are all notImplemented() stubs that are never called in the token-only flow.
+  console.log(`\nInitializing IssuerSparkWallet on ${network}...`);
 
-async function testSparkWallet(
-  mnemonic: string,
-  network: SparkNetwork
-): Promise<void> {
-  console.log(`\n── SparkWallet on ${network} ──────────────────────────────────`);
-
-  const walletOptions =
-    network === "MAINNET" ? WalletConfig.MAINNET : WalletConfig.REGTEST;
-  const { wallet, mnemonic: generatedMnemonic } = await SparkWallet.initialize({
-    mnemonicOrSeed: mnemonic,
-    options: walletOptions,
+  const { wallet } = await IssuerSparkWallet.initialize({
+    signer: signer as any,
+    options: {
+      network,
+      signerWithPreExistingKeys: true,
+      tokenSignatures: "SCHNORR",
+    },
   });
 
-  if (generatedMnemonic) {
-    console.log("New wallet generated – save this mnemonic:");
-    console.log(`  ${generatedMnemonic}`);
-    console.log("Set MNEMONIC in .env.local to reuse it.\n");
-  }
+  console.log(`✅ Authenticated to Spark SO`);
 
   const sparkAddress = await wallet.getSparkAddress();
-  const identityPubKey = await wallet.getIdentityPublicKey();
-  const { balance } = await wallet.getBalance();
+  const identityPubKey = await signer.getIdentityPublicKey();
+  console.log(`\nSpark address:       ${sparkAddress}`);
+  console.log(
+    `Identity public key: ${Buffer.from(identityPubKey).toString("hex")}`
+  );
 
-  console.log(`Spark address:       ${sparkAddress}`);
-  console.log(`Identity public key: ${identityPubKey}`);
-  console.log(`Balance:             ${balance.toLocaleString()} sats`);
+  // Step 1: CREATE token
+  console.log(
+    `\n── Step 1: CREATE token ────────────────────────────────────`
+  );
+  console.log(`  Name:       ${tokenName}`);
+  console.log(`  Ticker:     ${tokenTicker}`);
+  console.log(`  Decimals:   ${tokenDecimals}`);
+  console.log(`  Max supply: ${maxSupply.toLocaleString()}`);
 
-  const receiver = process.env.RECEIVER_SPARK_ADDRESS;
-  if (receiver) {
-    const amountSats = parseInt(process.env.TRANSFER_AMOUNT_SATS ?? "1000", 10);
-    console.log(`\nSending ${amountSats} sats to ${receiver} ...`);
-    const result = await wallet.transfer({ receiverSparkAddress: receiver, amountSats });
-    console.log(`Transfer complete: ${JSON.stringify(result)}`);
-  } else {
+  const createResult = await wallet.createToken({
+    tokenName,
+    tokenTicker,
+    decimals: tokenDecimals,
+    maxSupply,
+    isFreezable: false,
+  });
+  console.log(`✅ Token created`);
+  console.log(`   Result: ${JSON.stringify(createResult)}`);
+
+  // Get the token identifier (set by the SDK after creation)
+  const tokenIdentifier = await wallet.getIssuerTokenIdentifier();
+  if (!tokenIdentifier) {
+    throw new Error("Could not retrieve token identifier after creation");
+  }
+  console.log(`   Token identifier: ${tokenIdentifier}`);
+
+  // Step 2: MINT tokens
+  console.log(
+    `\n── Step 2: MINT ${mintAmount.toLocaleString()} tokens ────────────────────────────`
+  );
+
+  const mintTxId = await wallet.mintTokens(mintAmount);
+  console.log(`✅ Tokens minted`);
+  console.log(`   Tx ID: ${mintTxId}`);
+
+  // Check balance after mint
+  const { tokenBalances } = await wallet.getBalance();
+  for (const [tokenId, info] of tokenBalances) {
     console.log(
-      "\nSet RECEIVER_SPARK_ADDRESS (and optionally TRANSFER_AMOUNT_SATS) to test a Spark transfer."
+      `   Balance: ${info.ownedBalance} ${info.tokenMetadata?.tokenTicker ?? tokenId}`
     );
   }
 
+  // Step 3: TRANSFER tokens
+  console.log(
+    `\n── Step 3: TRANSFER ${transferAmount.toLocaleString()} tokens ────────────────────`
+  );
+  console.log(`   To: ${receiverSparkAddress}`);
+
+  const transferTxId = await wallet.transferTokens({
+    tokenIdentifier,
+    tokenAmount: transferAmount,
+    receiverSparkAddress,
+  });
+  console.log(`✅ Transfer broadcast`);
+  console.log(`   Tx ID: ${transferTxId}`);
+
+  // Check transaction status
+  const { tokenTransactionsWithStatus } =
+    await wallet.queryTokenTransactions({ tokenTransactionHashes: [transferTxId] });
+  if (tokenTransactionsWithStatus.length > 0) {
+    console.log(`   Status: ${tokenTransactionsWithStatus[0]!.status}`);
+  }
+
+  // Final balance
+  const finalBalance = await wallet.getBalance();
+  console.log(`\n── Final balances ─────────────────────────────────────────`);
+  for (const [tokenId, info] of finalBalance.tokenBalances) {
+    console.log(
+      `   ${info.tokenMetadata?.tokenTicker ?? tokenId}: ${info.ownedBalance}`
+    );
+  }
+
+  console.log(`\n✅ All Phase 1 operations succeeded with Turnkey signing!\n`);
   wallet.cleanupConnections();
 }
 
-// ---------------------------------------------------------------------------
-// main
-// ---------------------------------------------------------------------------
-
-async function main(): Promise<void> {
-  const network = (process.env.SPARK_NETWORK ?? "REGTEST") as SparkNetwork;
-  let mnemonic = process.env.MNEMONIC?.trim();
-
-  // Generate a fresh mnemonic if none provided
-  if (!mnemonic) {
-    const tmpSigner = new DefaultSparkSigner();
-    mnemonic = await tmpSigner.generateMnemonic();
-    console.log("No MNEMONIC set – generated a new one:");
-    console.log(`  ${mnemonic}`);
-    console.log("Add it to .env.local as MNEMONIC to reuse this wallet.\n");
-  }
-
-  console.log(`\nInitializing Spark signer on ${network}...`);
-  const signer = await initSigner(mnemonic);
-
-  const identityPubKey = await signer.getIdentityPublicKey();
-  const depositPubKey = await signer.getDepositSigningKey();
-  const sparkNetwork: Network =
-    network === "MAINNET" ? Network.MAINNET : Network.REGTEST;
-  const btcAddress = getP2TRAddressFromPublicKey(depositPubKey, sparkNetwork);
-
-  console.log(`\nIdentity public key: ${hex(identityPubKey)}`);
-  console.log(`Deposit public key:  ${hex(depositPubKey)}`);
-  console.log(`P2TR deposit addr:   ${btcAddress}`);
-
-  // Raw Schnorr comparison (identity key)
-  const testPayload = Buffer.from(
-    "spark-schnorr-test:"
-  );
-  await compareSchnorrSignatures(signer, testPayload);
-
-  // Bitcoin transaction signing (deposit key) – only when UTXO env vars are set
-  if (
-    process.env.UTXO_TXID &&
-    process.env.UTXO_VOUT &&
-    process.env.UTXO_VALUE &&
-    process.env.DESTINATION_ADDRESS
-  ) {
-    await signBitcoinTransaction(signer, depositPubKey);
-  } else {
-    console.log(
-      "\nSkipping Bitcoin tx signing – set UTXO_TXID, UTXO_VOUT, UTXO_VALUE, DESTINATION_ADDRESS to enable."
-    );
-  }
-
-  // Full Spark wallet test on regtest
-  await testSparkWallet(mnemonic, network);
-}
-
 main().catch((err) => {
-  console.error(err);
+  console.error("\n❌ Error:", err);
   process.exit(1);
 });
